requirements.txt
anthropic
backoff
beautifulsoup4
transformers
datasets
google-genai
openai
dotenv
ipython
pandas
torch
tqdm
sentence_transformers
